refactor(toolkit): introduce CodeInterpreterBuilder for provisioning
  Wrap the moved provisioning logic into CodeInterpreterBuilder, drop the
  underscore prefixes inside the private builder files, and point the
  builtin manager and tests at the builder. Lazy activation is stubbed
  with NotImplementedError pending the activator redesign.
