+++ unique_toolkit/unique_toolkit/agentic/tools/openai_builtin/code_interpreter/builder/__init__.py
@@ -1,0 +1,5 @@
+from unique_toolkit.agentic.tools.openai_builtin.code_interpreter.builder.builder import (
+    CodeInterpreterBuilder,
+)
+
+__all__ = ["CodeInterpreterBuilder"]
